+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## NEXT
 
-* New option to show Rinnai Touch module's status in logs & other minor improvements
+* New option to show module's status in logs & other minor improvements
 * JSONPath option to extract temperature from JSON payload for MQTT Current Temperature Subscription
+* Show module's fault status in the log and publish to MQTT
 
 ## 3.1.2 (2020-10-20)